# contracts/openapi/ninefold.openapi.yaml
openapi: 3.1.0
info:
  title: Ninefold Sudoku HTTP API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
...
  /replays/{matchId}/capabilities:
    post:
      operationId: createReplayCapability
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
      responses:
        '201':
          description: Replay read capability created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplayCapabilityResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /replays/{replayId}:
    get:
      operationId: getReplay
      parameters:
        - name: replayId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
      security:
        - replayCapability: []
      responses:
        '200':
          description: Immutable public replay projection
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteReplay
      parameters:
        - name: replayId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUIDv7'
      responses:
        '204':
          description: Replay access deleted
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    replayCapability:
      type: http
      scheme: bearer
  schemas:
    UUIDv7:
      type: string
      pattern: '^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$'
    SafeInteger:
      type: integer
      format: int64
      minimum: 0
      maximum: 9007199254740991
...
    ReplayCapabilityResponse:
      type: object
      additionalProperties: false
      required: [replayId, capability, shareUrl, expiresAt]
      properties:
        replayId:
          $ref: '#/components/schemas/UUIDv7'
        capability:
          type: string
          minLength: 43
          maxLength: 43
        shareUrl:
          type: string
          minLength: 1
          maxLength: 512
        expiresAt:
          $ref: '#/components/schemas/SafeInteger'
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, messageKey, requestId, details]
          properties:
            code:
              type: string
              pattern: '^[A-Z][A-Z0-9_]{1,63}$'
            messageKey:
              type: string
              pattern: "^error\\.[a-z0-9_]+$"
            requestId:
              type: string
              maxLength: 36
            retryable:
              type: boolean
            details:
              type: object
              propertyNames:
                pattern: '^(?!.*(?:[Tt]oken|[Ss]ecret|[Cc]ookie|[Cc]redential|[Cc]apability|[Ss]olution)).+$'
              additionalProperties:
                oneOf:
                  - type: string
                  - type: boolean
                  - type: integer
                    format: int64
                    minimum: 0
                    maximum: 9007199254740991
                  - type: 'null'
  responses:
...
    NotFound:
      description: Room or resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...